Untested .UF2
Should be the better behaved interpreter.

Connect to it on both USB (at /dev/ttyACM0) as well
as on the USART pins (upper left corner).

Mirrors on both, with individually chosen output
(the mirroring of any given utterance is entirely
arbitrary; the mirroring goes on in parallel, for
each type of utterance).

	modified:   07-interpreter.d/pico-examples/build/interpreter-a/interpreter-a.uf2

On branch dvlp-aa-